ENH: Separate plugins per namespace
