+++ loktar/src/tests/test_plugins_whl.py
@@ -1,9 +1,9 @@
 import pytest
 
 from loktar.exceptions import CIBuildPackageFail
-from loktar.plugins.whl import PypicloudClient
-from loktar.plugins.whl import run
-from loktar.plugins.whl import Whl
+from loktar.plugins.artifact.whl import PypicloudClient
+from loktar.plugins.artifact.whl import run
+from loktar.plugins.artifact.whl import Whl
 
 
 @pytest.fixture()
@@ -20,9 +20,9 @@ def os_environ(mocker):
 @pytest.mark.parametrize("mode", ["master", "foobar"])
 @pytest.mark.parametrize("remote", [True, False])
 def test_plugins_whl(mocker, mode, remote):
-    mocker.patch("loktar.plugins.whl.PypicloudClient")
-    mocker.patch("loktar.plugins.whl.exe")
-    mocker.patch("loktar.plugins.whl.exec_command_with_retry")
+    mocker.patch("loktar.plugins.artifact.whl.PypicloudClient")
+    mocker.patch("loktar.plugins.artifact.whl.exe")
+    mocker.patch("loktar.plugins.artifact.whl.exec_command_with_retry")
     mocker.patch("loktar.plugin.exe")
     run({"pkg_name": "foobar", "mode": mode, "package_location": "/tmp"}, remote)
 
@@ -30,9 +30,9 @@ def test_plugins_whl(mocker, mode, remote):
 @pytest.mark.parametrize("mode", ["master", "foobar"])
 @pytest.mark.parametrize("remote", [True, False])
 def test_plugins_whl_fail_on_update_version(mocker, mode, remote):
-    mocker.patch("loktar.plugins.whl.PypicloudClient")
-    mocker.patch("loktar.plugins.whl.exe", return_value=False)
-    mocker.patch("loktar.plugins.whl.exec_command_with_retry")
+    mocker.patch("loktar.plugins.artifact.whl.PypicloudClient")
+    mocker.patch("loktar.plugins.artifact.whl.exe", return_value=False)
+    mocker.patch("loktar.plugins.artifact.whl.exec_command_with_retry")
     mocker.patch("loktar.plugin.exe")
     with pytest.raises(CIBuildPackageFail):
         run({"pkg_name": "foobar", "mode": mode, "package_location": "/tmp"}, remote)
@@ -41,9 +41,9 @@ def test_plugins_whl_fail_on_update_version(mocker, mode, remote):
 @pytest.mark.parametrize("mode", ["master", "foobar"])
 @pytest.mark.parametrize("remote", [True, False])
 def test_plugins_whl_fail_on_release(mocker, mode, remote):
-    mocker.patch("loktar.plugins.whl.PypicloudClient")
-    mocker.patch("loktar.plugins.whl.exe")
-    mocker.patch("loktar.plugins.whl.exec_command_with_retry", return_value=False)
+    mocker.patch("loktar.plugins.artifact.whl.PypicloudClient")
+    mocker.patch("loktar.plugins.artifact.whl.exe")
+    mocker.patch("loktar.plugins.artifact.whl.exec_command_with_retry", return_value=False)
     mocker.patch("loktar.plugin.exe")
     with pytest.raises(CIBuildPackageFail):
         run({"pkg_name": "foobar", "mode": mode, "package_location": "/tmp"}, remote)
